requirements.txt
jieba~=0.42.1
numpy~=1.26.4
pillow~=10.4.0
paddleocr~=2.9.1
pathlib~=1.0.1
opencv-python~=4.10.0.84
torch~=2.5.1
ultralytics~=8.3.28
yaml~=0.2.5
pyyaml~=6.0.1
torchvision~=0.20.1
matplotlib~=3.9.2
pandas~=2.2.2
seaborn~=0.13.2
scipy~=1.13.1
requests~=2.32.3
tqdm~=4.67.0
psutil~=5.9.0
thop~=0.1.1-2209072238